refactor: move purgomalum client under the 'common/infra' package
